ai-youtube-automation/video_editing_api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
moviepy==1.0.3
opencv-python==4.10.0.84
pytesseract==0.3.13
Pillow==10.4.0
pyttsx3==2.90
python-multipart==0.0.9